+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Removed development ribbon
 * Updated angular2-material to 2.0.0-alpha.5-2
 * Added angular2-material palette to main.scss
+* Added sidenav navigation
 
 ### 1.6.0 (26.05.2016)
 * Updated node to 6.2.0