+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Changelog
 
-### 1.7.0 (unreleased)
+### 1.7.0 (08.06.2016)
 * Added npm-check-updates as dev dependency
 * Removed kitchensink page
 * Updated dependencies